webhook/requirements.txt
flask
mysql-connector-python
gunicorn
pyTelegramBotAPI
python-dotenv
datetime
pytz
